Added Functional model layers support for GAN
Functional layer support helps to make models using generator and
discriminator with multiple inputs and outputs to effectively make GAN
flow. It helps to build ClassifierGAN, StyleGAN, CycleGAN and a lot more
architectures in ECL to work with data efficiently. Test file with CGAN
will be added in further commits.